download+upload -> new function "ck transfer" for moving entries around

+++ module/misc/module.py
@@ -1468,12 +1468,14 @@ def replace_strings_in_file(i):
     return r
 
 
-def download(i):
+def transfer(i):
     """
     Input:  {
                 xcids[]             - (remote) entries to be copied to a local repository, made of repo_uoa:module_uoa:data_uoa
 
-                (local_repo_uoa)    - where to store the copies locally ('local' by default)
+                (source_repo_uoa)   - in case the source is remote, the specific repository on the source (optional - only needed for disambiguation)
+                (target_server_uoa) - the optional target remote server (mapped as a local repository)
+                (target_repo_uoa)   - the repository to store the entries on target server ('local' by default)
             }
 
     Output: {
@@ -1485,67 +1487,77 @@ def download(i):
 
     import os
 
-    o               = i.get('out','')
-    local_repo_uoa  = i.get('local_repo_uoa', 'local')
-    source_addrs    = i.get('xcids', [])       # can unshift the "default entry" into this list when/if moving this method to the kernel
+    o                   = i.get('out','')
+    source_addrs        = i.get('xcids', [])       # can unshift the "default entry" into this list when/if moving this method to the kernel
 
+    source_repo_uoa     = i.get('source_repo_uoa')
+    target_server_uoa   = i.get('target_server_uoa')
+    target_repo_uoa     = i.get('target_repo_uoa', 'local')
+
+    source_addr_ext     = {'remote_repo_uoa': source_repo_uoa} if source_repo_uoa else {}
+    target_addr_ext     = {'repo_uoa': target_server_uoa, 'remote_repo_uoa': target_repo_uoa} if target_server_uoa else {'repo_uoa': target_repo_uoa}
 
     if len(source_addrs)==0:
-        return {'return':1, 'error': 'Need a non-empty list of remote CID addresses to download the entries from'}
+        return {'return':1, 'error': 'Need a non-empty list of source CID addresses'}
 
     for addr in source_addrs:
-        server_repo_uoa    = addr['repo_uoa']
-        if (not server_repo_uoa) or server_repo_uoa=='local':
-            return {'return':2, 'error': 'All entries to be copied have to be remote ({})'.format(addr)}
+        repo_uoa    = addr.get('repo_uoa')
+        if (not repo_uoa) and (not target_server_uoa):
+            return {'return':2, 'error': 'Cannot copy the entries locally (since IDs are to be preserved)'}
 
         load_adict = {  'action':           'load',
-                        'repo_uoa':         server_repo_uoa,
                         'module_uoa':       addr['module_uoa'],
                         'data_uoa':         addr['data_uoa'],
         }
+        if repo_uoa: load_adict.update( {'repo_uoa': repo_uoa} )
+        load_adict.update( source_addr_ext )
         r=ck.access( load_adict )
         if r['return']>0: return r
 
-        meta_dict       = r['dict']
-        data_uid        = r['data_uid']
-        remote_repo_uoa = r['repo_uoa']     # this is where we originally found it...
+        meta_dict           = r['dict']
+        data_uid            = r['data_uid']
+        original_repo_uoa   = r['repo_uoa']     # this is where we originally found it...
 
         add_adict = {   'action':           'add',
                         'common_func':      'yes',
-                        'repo_uoa':         local_repo_uoa,
                         'module_uoa':       addr['module_uoa'],
                         'data_uoa':         addr['data_uoa'],
                         'dict':             meta_dict,              # copying meta data
                         'data_uid':         data_uid,               # copying the original data_uid
         }
+        add_adict.update( target_addr_ext )
         r=ck.access( add_adict )
         if r['return']>0: return r
 
         pull_adict = {  'action':           'pull',
-                        'repo_uoa':         server_repo_uoa,
                         'module_uoa':       addr['module_uoa'],
                         'data_uoa':         addr['data_uoa'],
                         'archive':          'yes',
         }
+        if repo_uoa: pull_adict.update( {'repo_uoa': repo_uoa} )
+        pull_adict.update( source_addr_ext )
         r=ck.access( pull_adict )
         if r['return']>0: return r
 
         zip_name = ck.cfg['default_archive_name']
 
         push_adict = {  'action':           'push',
-                        'repo_uoa':         local_repo_uoa,
                         'module_uoa':       addr['module_uoa'],
                         'data_uoa':         addr['data_uoa'],
                         'archive':          'yes',
                         'filename':         zip_name,
         }
+        push_adict.update( target_addr_ext )
         r=ck.access( push_adict )
         if r['return']>0: return r
 
-        os.remove(zip_name)     # to avoid clashing with the next one
+        if os.path.isfile(zip_name):
+            os.remove(zip_name)     # to avoid clashing with the next one
 
         if o=='con':
-            ck.out('{}[{}]:{}:{} -> {}:{}:{}' \
-                .format(server_repo_uoa, remote_repo_uoa, addr['module_uoa'], addr['data_uoa'], local_repo_uoa,addr['module_uoa'], addr['data_uoa']))
+            source_addr = original_repo_uoa if repo_uoa in (original_repo_uoa, None) else '{}/{}'.format(repo_uoa, original_repo_uoa)
+            target_addr = '{}/{}'.format(target_server_uoa, target_repo_uoa) if target_server_uoa else target_repo_uoa
+            ck.out('{}:{}:{} -> {}:{}:{}' \
+                .format(source_addr, addr['module_uoa'], addr['data_uoa'], target_addr, addr['module_uoa'], addr['data_uoa']))
 
     return {'return':0}
